refactor some file position
