+++ src/storage/src/hummock/sstable/delete_range_aggregator.rs
@@ -18,7 +18,8 @@ use std::sync::Arc;
 
 use risingwave_hummock_sdk::HummockEpoch;
 
-use crate::hummock::DeleteRangeTombstone;
+use super::DeleteRangeTombstone;
+use crate::hummock::iterator::DeleteRangeIterator;
 
 pub struct SortedBoundary {
     sequence: HummockEpoch,
@@ -136,18 +137,6 @@ impl DeleteRangeAggregator {
         }
         delete_ranges
     }
-}
-
-pub trait DeleteRangeIterator {
-    fn start_user_key(&self) -> &[u8];
-    fn end_user_key(&self) -> &[u8];
-    fn current_epoch(&self) -> HummockEpoch;
-    fn next(&mut self);
-    fn seek_to_first(&mut self);
-    fn seek(&mut self, _target_key: &[u8]) {
-        unimplemented!("Support seek operation");
-    }
-    fn valid(&self) -> bool;
 }
 
 pub struct SingleDeleteRangeIterator {
